Recomboulating stuff
Firstly, trying to remake a more organized directory for teaching notes.
Secondly, added a bit of javascript to the beginning of the endowment
graph example to make it work a bit better on mobile.

From testing on my laptop, it seems like the equations are more
readable but the graph gets a bit distorted.