Fiddled with colors and Latex stylin
I used the inspection tool in chrome to find the class names of the
mathjax objects and the equation tag in particular. Then I used css to
override the default styling and add custom colors.

I also enabled the auto bold for Mathjax 2.7

After a bunch of experimentation, I really like the speed of MathJax 3,
but there are still some rendering issues.
Sometimes overlies are drawn in the wrong place.
Sometimes equations are scrunched too tightly together.
In safari, big chunks of text are missing. etc.