Changed formatting on Interactive Example
axis size and chart height are now set to "100%".
The chart height is a percentage of the chart width, so I removed the
chart width from the styling above.
For the axes, they default to a percentage of the total chart size. But
the boxes don't quite line up because the numbers on the y axis are
wider than the nums on the x axis are tall.

Also fixed a few assorted typos.