Homepage Tweaks
- Middle Name included in title.
- Image occurs after title.
- Contact Info Header h2 style modified to fix the issue where certain screen widths cause wonkiness.

For the h2:
The header's line was starting after the image's padding, halfway across the page, then looping around the left. I only want the header to start halfway across the page if the entire header is like that.

`clear:both` made things look proper on mobile, but not on desktop.
`display:inline-block` solved the issue but added extra margins. (Or maybe made the margins not overlap with adjacent elements.
So I also removed the margins on the heading.