Update AD_AS, The Labor Market, and the Phillips Curve - EconGraphs.html
Actual change: just switched to a different version in the test file.

More more importantly: The hosted version works great. The local CORS things is baked in as a security feature apparently.

https://stackoverflow.com/questions/48362093/cors-request-blocked-in-locally-opened-html-file

And as always with this stuff, the fix seems barely tailored to the bug. ah, well. The linked stackoverflow page has some ways to do local hosting (?!!Fff?!) to get around the issue.