330 HW1,Q1 - fiddle with styling on graph
I changed the column labels to be more concise but worse for labels, and switched to using explicit individual lineplot calls.

But this made the linestyle of the second graph revert to solid. Which confuses me because on its own, the df plotting wasn't dashed, but the df plotting with seaborn was?  But the individual seaborn plotting wasn't?
And why aren't the dash patterns the same?

Ah well, none of this matters.